+++ packages/placement-ordering/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.0.1...@pie-element/placement-ordering@5.1.0) (2019-05-16)
+
+
+### Features
+
+* Sending rationale for `instructor` role only in 'view' or 'evaluate' mode. ([ea90849](https://github.com/pie-framework/pie-elements/commit/ea90849))
+* **placement-ordering:** added rationale support. ([d6df005](https://github.com/pie-framework/pie-elements/commit/d6df005))
+
+
+
+
+
 ## [5.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.0.0...@pie-element/placement-ordering@5.0.1) (2019-05-08)
 
 **Note:** Version bump only for package @pie-element/placement-ordering